+++ src/app/dashboard/page.js
@@ -1,6 +1,5 @@
 "use client";
 import Table from "@/components/dashboard/Table";
-import FilterBar from "@/components/FilterBar";
 import { useState } from "react";
 
 const Page = () => {
@@ -43,7 +42,6 @@ const Page = () => {
 
   return (
     <div className="w-full flex flex-col justify-center items-center">
-      {<FilterBar />}
       <Table
         objects={objects}
         setObjects={setObjects}
